Give hints about failure to converge (#3)
This defaults to a short timeout, and provides hints about how
to adjust settings to control convergence criteria. It also returns
the convergence status, exporting `OPTIMAL` so that users can
test whether the algorithm converged.

This also adds docstrings and moves the `parse_input` docstring to
the function (so that it is always available).